## What Changed
- Closed #1544 

## Evidence
### Integration-Test
https://github.com/Yamato-Security/hayabusa/actions/runs/12831409129

I would appreciate it if you could check it out when you have time🙏